Renamed components to PascalCase
