+++ src/js/components/Char.js
