+++ src/js/components/Hint.js
