+++ src/js/components/Logo.js
