+++ src/js/components/ProgressBar.js
